Add new Automattic color palette
This is a first iteration to be used primarily in Automattic for Agencies (A4A).

While it's a decent starting point, we'd like to keep improving it as we apply more of it to real interfaces, and uncover its limitations and quirks.

The primary brand color differs from what's been used in other materials a bit, but the new #24A3E0 color is very close anyhow that it could be implemented with no discernible difference.